requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==25.4.0
bcrypt==5.0.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
Faker==40.1.2
greenlet==3.3.0
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
jsonpath==0.82.2
loguru==0.7.3
lxml==6.0.2
more-itertools==10.8.0
packaging==25.0
paramiko==4.0.0
pathvalidate==3.3.1
playwright==1.57.0
pluggy==1.6.0
premailer==3.10.0
pycparser==2.23
pycryptodome==3.23.0
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.2
PyNaCl==1.6.2
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
pytest-rerunfailures==16.1
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
schedule==1.2.2
sshtunnel==0.4.0
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
yagmail==0.15.293
